restructure parsers
