+++ apps/github-receiver/src/core/parsers/base.parser.ts
@@ -1,0 +1,3 @@
+export abstract class GitHubParser<T, R> {
+  abstract parse(payload: T): R;
+}
