+++ apps/github-receiver/src/core/parsers/push.parser.ts
@@ -1,0 +1,25 @@
+import { PushEventData } from '@codeheroes/providers';
+import { PushEvent } from '../interfaces/github.interfaces';
+import { GitHubParser } from './base.parser';
+
+export class PushEventParser extends GitHubParser<PushEvent, PushEventData> {
+  parse(payload: PushEvent): PushEventData {
+    return {
+      repository: {
+        id: payload.repository.id.toString(),
+        name: payload.repository.name,
+        owner: payload.repository.owner.login,
+        ownerType: payload.repository.owner.type,
+      },
+      metrics: {
+        commits: payload.commits.length,
+      },
+      branch: payload.ref,
+      lastCommitMessage: payload.head_commit?.message || null,
+      sender: {
+        id: payload.sender.id.toString(),
+        login: payload.sender.login,
+      },
+    };
+  }
+}
